refactor: block-tracker into separate files
